+++ snapshots/snapshots/src/main/kotlin/com/emergetools/snapshots/EmergeSnapshots.kt
@@ -1,7 +1,7 @@
 package com.emergetools.snapshots
 
 import android.graphics.Bitmap
-import com.emergetools.snapshots.shared.ComposePreviewSnapshotConfig
+import com.emergetools.snapshots.models.ComposePreviewSnapshotConfig
 import org.junit.rules.TestRule
 import org.junit.runner.Description
 import org.junit.runners.model.Statement