+++ snapshots/snapshots/src/main/kotlin/com/emergetools/snapshots/compose/EmergeComposeSnapshotReflectiveParameterizedInvoker.kt
@@ -6,8 +6,8 @@ import androidx.compose.ui.tooling.PreviewActivity
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.platform.app.InstrumentationRegistry
 import com.emergetools.snapshots.EmergeSnapshots
-import com.emergetools.snapshots.shared.ComposePreviewSnapshotConfig
-import com.emergetools.snapshots.shared.ComposeSnapshots
+import com.emergetools.snapshots.models.ComposePreviewSnapshotConfig
+import com.emergetools.snapshots.models.ComposeSnapshots
 import com.emergetools.snapshots.util.Profiler
 import kotlinx.serialization.json.Json
 import org.junit.Rule