+++ snapshots/snapshots/src/main/kotlin/com/emergetools/snapshots/models/ComposePreviewSnapshotConfig.kt
@@ -1,4 +1,4 @@
-package com.emergetools.snapshots.shared
+package com.emergetools.snapshots.models
 
 import kotlinx.serialization.Serializable