+++ snapshots/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to snapshots & snapshots-processor will be documented in thi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Remove `snapshots-shared` module & dependency.
+
 ## 1.3.4 - 2025-03-21
 
 - Add `@EmergeSnapshotConfig` annotation for setting `ignore` and